Chat message api done #9

+++ src/app/modules/message/Message.service.ts
@@ -1,11 +1,14 @@
 import { StatusCodes } from 'http-status-codes';
 import ServerError from '../../../errors/ServerError';
-import { prisma } from '../../../utils/db';
+import { type Prisma, prisma } from '../../../utils/db';
 import type {
   TCreateMessageArgs,
   TDeleteMessageArgs,
+  TGetChatMessagesArgs,
 } from './Message.interface';
 import { deleteFiles } from '../../middlewares/capture';
+import { messageSearchableFields } from './Message.constant';
+import type { TPagination } from '../../../utils/server/serveResponse';
 
 /**
  * All message related services
@@ -15,15 +18,36 @@ export const MessageServices = {
    * Create new message
    */
   async createMessage(payload: TCreateMessageArgs) {
-    return prisma.message.create({
-      data: payload,
-      include: {
-        chat: {
-          select: {
-            user_ids: true,
+    return prisma.$transaction(async tx => {
+      //? update chat timestamp
+      tx.chat.update({
+        where: { id: payload.chat_id },
+        data: { timestamp: new Date() },
+      });
+
+      //? create message
+      return tx.message.create({
+        data: {
+          ...payload,
+          seen_by: {
+            connect: {
+              id: payload.user_id,
+            },
           },
         },
-      },
+        include: {
+          chat: {
+            select: {
+              user_ids: true,
+            },
+          },
+          seen_by: {
+            select: {
+              avatar: true,
+            },
+          },
+        },
+      });
     });
   },
 
@@ -33,7 +57,7 @@ export const MessageServices = {
   async deleteMessage({ message_id, user_id }: TDeleteMessageArgs) {
     const message = await prisma.message.findUnique({
       where: { id: message_id },
-      select: { user_id: true, media_urls: true },
+      select: { user_id: true, media_urls: true, isDeleted: true },
     });
 
     //? ensure that user has permission to delete message
@@ -44,17 +68,80 @@ export const MessageServices = {
       );
     }
 
+    if (message.isDeleted) {
+      throw new ServerError(StatusCodes.BAD_REQUEST, 'Message already deleted');
+    }
+
     await deleteFiles(message.media_urls);
 
-    return prisma.message.delete({
+    return prisma.message.update({
       where: { id: message_id },
+      data: {
+        media_urls: [],
+        text: '𝒹𝑒𝓁𝑒𝓉𝑒𝒹 𝓂𝑒𝓈𝓈𝒶𝑔𝑒',
+        isDeleted: true,
+      },
       select: {
         chat: {
+          select: { id: true, user_ids: true },
+        },
+      },
+    });
+  },
+
+  /**
+   * Get chat messages
+   */
+  async getChatMessages({
+    chat_id,
+    limit,
+    page,
+    search,
+  }: TGetChatMessagesArgs) {
+    const messageWhere: Prisma.MessageWhereInput = {
+      chat_id,
+    };
+
+    if (search) {
+      messageWhere.OR = messageSearchableFields.map(field => ({
+        [field]: {
+          contains: search,
+          mode: 'insensitive',
+        },
+      }));
+    }
+
+    const messages = await prisma.message.findMany({
+      where: messageWhere,
+      skip: (page - 1) * limit,
+      take: limit,
+      orderBy: {
+        created_at: 'desc',
+      },
+      include: {
+        seen_by: {
           select: {
-            user_ids: true,
+            avatar: true,
           },
         },
       },
     });
+
+    const total = await prisma.message.count({ where: messageWhere });
+
+    return {
+      meta: {
+        pagination: {
+          page,
+          limit,
+          total,
+          totalPages: Math.ceil(total / limit),
+        } satisfies TPagination,
+      },
+      messages: messages.map(({ seen_by, ...message }) => ({
+        ...message,
+        seen_by: seen_by.map(user => user.avatar),
+      })),
+    };
   },
 };
